Refactor file structure co-location pattern with index folders
