--- src/components/App.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import HomePage from './home-page';
-import './App.css';
-
-const App = () => {
-  return <HomePage />
-}
-
-export default App;
