+++ src/components/Footer/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+// Define functional component for Footer
+const Footer: React.FC = () => {
+  return (
+    <footer>
+      <p>&copy; 2023 My Home Page</p>
+    </footer>
+  );
+};
+
+export default Footer;
