+++ src/components/index.tsx
@@ -1,0 +1,3 @@
+export { default as App } from './App';
+export { default as Footer } from './Footer';
+export { default as Header } from "./Header";
