+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './components/App';
+import { App } from './components';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(
