+++ src/pages/AboutPage/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+// Define functional component for AboutPage
+const AboutPage: React.FC = () => {
+  return (
+    <main>
+      <h2>About Us</h2>
+      <p>We are a marketing and web development agency based in the United States.</p>
+    </main>
+  );
+};
+
+export default AboutPage;
