+++ src/pages/ContactPage/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+// Define functional component for ContactPage
+const ContactPage: React.FC = () => {
+  return (
+    <main>
+      <h2>Contact Us</h2>
+      <p>You can reach us at info@brandgeneering.com.</p>
+    </main>
+  );
+};
+
+export default ContactPage;
