+++ src/pages/MainContent/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+// Define functional component for MainContent
+const MainContent: React.FC = () => {
+  return (
+    <main>
+      <h2>BrandGeneering Consulting</h2>
+      <h3>Marketing, Strategy, and Web Applications</h3>
+    </main>
+  );
+};
+
+export default MainContent;
