+++ src/pages/index.tsx
@@ -1,0 +1,4 @@
+export { default as AboutPage } from './AboutPage';
+export { default as ContactPage } from './ContactPage';
+export { default as HomePage } from './HomePage';
+export { default as MainContent } from "./MainContent";
